[feature/converter] Code cleanup and fixes. Code cleaned up. Moved to use of language files.
