+++ phpBB/adm/style/converter_main.html
@@ -3,7 +3,7 @@
 	<h1>{TITLE}</h1>
 	<p>{CONTENT}</p>
 <form id="install_install" method="POST" action="{U_ACTION}">
-	<fieldset class="Source Database Details">
+	<fieldset class="db_details">
 		<legend>Database Details</legend>
 		<dl>
 			<dt>
@@ -41,4 +41,4 @@
 </form>
 <div><a href="{U_LINK}">Next</a></div>
 {{ include('installer_footer.html') }}
-<script src="/assets/javascript/converter.js"></script>
+<!-- INCLUDEJS {T_ASSETS_PATH}/javascript/converter.js -->
