+++ phpBB/adm/style/converter_process.html
@@ -13,4 +13,5 @@
 
 
 {{ include('installer_footer.html') }}
-<script src="/assets/javascript/converter.js"></script>
+<!-- INCLUDEJS {T_ASSETS_PATH}/javascript/converter.js -->
+
